fix test

+++ core/payments/reservation/reservation_ledger.go
@@ -26,7 +26,7 @@ type ReservationLedger struct {
 // Creates a new reservation ledger, which represents the reservation of a single user with a [LeakyBucket]
 func NewReservationLedger(
 	config ReservationLedgerConfig,
-	// now should be from a source that includes monotonic timestamp for best results
+// now should be from a source that includes monotonic timestamp for best results
 	now time.Time,
 ) (*ReservationLedger, error) {
 	leakyBucket, err := ratelimit.NewLeakyBucket(
@@ -61,15 +61,15 @@ func NewReservationLedger(
 // If the bucket doesn't have enough capacity to accommodate the fill, symbolCount IS NOT added to the bucket, i.e. a
 // failed debit doesn't count against the meter.
 func (rl *ReservationLedger) Debit(
-	// now should be from a source that includes monotonic timestamp for best results.
-	// This a local time from the perspective of the entity that owns this ledger instance, to be used with the local
-	// leaky bucket: it should NOT be sourced from the PaymentHeader
+// now should be from a source that includes monotonic timestamp for best results.
+// This a local time from the perspective of the entity that owns this ledger instance, to be used with the local
+// leaky bucket: it should NOT be sourced from the PaymentHeader
 	now time.Time,
-	// the timestamp included, or planned to be included, in the PaymentHeader
+// the timestamp included, or planned to be included, in the PaymentHeader
 	dispersalTime time.Time,
-	// the number of symbols to debit
+// the number of symbols to debit
 	symbolCount uint32,
-	// the quorums being dispersed to
+// the quorums being dispersed to
 	quorums []core.QuorumID,
 ) (bool, float64, error) {
 
@@ -168,7 +168,7 @@ func (rl *ReservationLedger) UpdateReservation(newReservation *Reservation, now 
 	previousFillLevel := rl.leakyBucket.CheckFillLevel(now)
 
 	newLeakyBucket, err := ratelimit.NewLeakyBucket(
-		newConfig.reservation.symbolsPerSecond,
+		float64(newConfig.reservation.symbolsPerSecond),
 		newConfig.bucketCapacityDuration,
 		false, // fill level is explicitly set below
 		newConfig.overfillBehavior,
